wip: more refactor
